Remove first-redirect hack
Instead of having showIndexPage() redirect us, let the redirects
themselves take care of calling it.

This removes the odd piece of code which decided whether to replace the
url or not. Instead, have the '/' path be a different route which
replaces the url with the latest version and triggers that route.