Wait for the data to load before rendering
When setting the version, that can trigger an asynchronous request to
download the json file. We were assuming the new data was set when the
function returned, which meant on a new load, we were not rendering the
right version, but whatever happened to be loaded before.

Use a callback to render the views only once the new data has been
downloaded and set (or immediately if we're already in the right
version).